# core/src/main/resources/openapi/openapi-spec.yml
openapi: 3.0.3
info:
  title: Task by Artur Jarosz
  description: "TASK app openAPI specification"
  version: 0.6.0-SNAPSHOT
...
paths:
  /architects:
    post:
      tags:
        - architect
      summary: Create a new architect
      operationId: createArchitect
      requestBody:
        description: New architect data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Architect"
      responses:
        201:
          description: Architect successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Architect"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
    get:
      tags:
        - architect
      summary: Get list of architects
      operationId: getArchitects
      responses:
        201:
          description: Architect successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Architects"
  /architects/{architectId}:
    delete:
      tags:
        - architect
      summary: Delete an architect with given architectId
      operationId: deleteArchitect
      parameters:
        - name: architectId
          in: path
          description: "A unique identifier of the architect"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Successful operation
          content: { }
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
    get:
      tags:
        - architect
      summary: Get an architect with given architectId
      operationId: getArchitect
      parameters:
        - name: architectId
          in: path
          description: "A unique identifier of the architect"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Architect"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
    put:
      tags:
        - architect
      summary: Update an architect with given architectId
      operationId: updateArchitect
      parameters:
        - name: architectId
          in: path
          description: "A unique identifier of the architect"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: Updated data of the architect
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Architect"
      responses:
        200:
          description: Architect successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Architect"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
...
  /contracts/{contractId}/reject:
    post:
      tags:
        - contract
      summary: Reject contract offer
      operationId: rejectOffer
      parameters:
        - name: contractId
          in: path
          description: "A unique identifier of the contract"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Contract rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
  /contracts/{contractId}/new-offer:
    post:
      tags:
        - contract
      summary: Make a new offer
      operationId: makeNewOffer
      parameters:
        - name: contractId
          in: path
          description: "A unique identifier of the contract"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: New offer for contract made
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Contract"
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
  /contracts/{contractId}/accept-offer:
    post:
      tags:
        - contract
      summary: Accepts contract offer
      operationId: acceptOffer
      parameters:
        - name: contractId
          in: path
          description: "A unique identifier of the contract"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Offer for the contract accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
...
  /contracts/{contractId}/terminate:
    post:
      tags:
        - contract
      summary: Terminate the contract
      operationId: terminate
      parameters:
        - name: contractId
          in: path
          description: "A unique identifier of the contract"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: Data of contract to terminate
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Contract"
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
...
  /contracts/{contractId}/complete:
    post:
      tags:
        - contract
      summary: Complete the contract
      operationId: complete
      parameters:
        - name: contractId
          in: path
          description: "A unique identifier of the contract"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: Data of contract to complete
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Contract"
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
...
  /projects/{projectId}/contractor-jobs:
    post:
      tags:
        - contractorJob
      summary: Create a new contractor job
      operationId: createContractorJob
      parameters:
        - name: projectId
          in: path
          description: "A unique identifier of the project"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: New contractor job data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractorJob"
      responses:
        201:
          description: Contractor job successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractorJob"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
  /projects/{projectId}/contractor-jobs/{contractorJobId}:
    delete:
      tags:
        - contractorJob
      summary: Delete a contractor job with given contractorJobId
      operationId: deleteContractorJob
      parameters:
        - name: projectId
          in: path
          description: "A unique identifier of the project"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: contractorJobId
          in: path
          description: "A unique identifier of the contractor job"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Contractor job successfully deleted
          content: { }
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
    put:
      tags:
        - contractorJob
      summary: Update a contractor job
      operationId: updateContractorJob
      parameters:
        - name: projectId
          in: path
          description: "A unique identifier of the project"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: contractorJobId
          in: path
          description: "A unique identifier of the contractor job"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: Updated contractor job data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractorJob"
      responses:
        200:
          description: Contractor job successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractorJob"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
    get:
      tags:
        - contractorJob
      summary: Get a contractor job
      operationId: getContractorJob
      parameters:
        - name: projectId
          in: path
          description: "A unique identifier of the project"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: contractorJobId
          in: path
          description: "A unique identifier of the contractor job"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractorJob"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
...
components:
  schemas:
    Architect:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
    Architects:
      type: array
      items:
        $ref: "#/components/schemas/Architect"
...
    Contract:
      type: object
      properties:
        id:
          type: integer
          format: int64
        offerValue:
          type: number
          format: double
        signingDate:
          type: string
          format: date
        deadline:
          type: string
          format: date
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/ContractStatus"
        nextStatuses:
          items:
            $ref: "#/components/schemas/ContractStatus"
    ContractStatus:
      type: string
      enum:
        - OFFER
        - REJECTED
        - ACCEPTED
        - SIGNED
        - TERMINATED
        - COMPLETED
...
    ContractorJob:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        value:
          type: number
        contractorId:
          type: integer
          format: int64
        note:
          type: string
        hasInvoice:
          type: boolean
        payable:
          type: boolean
        paid:
          type: boolean
...
    ValidatorError:
      type: object
      properties:
        message:
          type: string
...